@@ -6,18 +6,18 @@
 		"@codingame/monaco-jsonrpc": "^0.3.1",
 		"@codingame/monaco-languageclient": "^0.17.3",
 		"express": "^4.15.2",
-		"file-loader": "^4.3.0",
-		"monaco-editor": "^0.30.1",
+		"file-loader": "^6.2.0",
+		"monaco-editor": "^0.31.1",
 		"normalize-url": "^7.0.3",
-		"reconnecting-websocket": "^3.2.2",
-		"request-light": "^0.2.2",
+		"reconnecting-websocket": "^4.4.0",
+		"request-light": "^0.5.7",
 		"source-map": "^0.7.3",
 		"vscode-json-languageservice": "^4.1.9",
 		"vscode-jsonrpc": "^6.0.0",
 		"vscode-languageserver": "7.0.0",
 		"vscode-uri": "^3.0.2",
-		"webpack": "^4.44.2",
-		"ws": "^5.0.0"
+		"webpack": "^5.68.0",
+		"ws": "^8.4.2"
 	},
 	"scripts": {
 		"prepare": "npm run clean && npm run build",
@@ -32,18 +32,18 @@
 	"devDependencies": {
 		"@types/express": "^4.0.35",
 		"@types/glob": "^7.2.0",
-		"@types/node": "^13.1.7",
-		"@types/normalize-url": "^1.9.1",
-		"@types/vscode": "1.52.0",
-		"@types/ws": "0.0.39",
-		"css-loader": "^0.28.11",
+		"@types/node": "^17.0.14",
+		"@types/normalize-url": "^4.2.0",
+		"@types/vscode": "1.63.2",
+		"@types/ws": "8.2.2",
+		"css-loader": "^6.6.0",
 		"monaco-editor-core": "^0.31.1",
 		"os-browserify": "^0.3.0",
-		"rimraf": "^2.6.2",
-		"source-map-loader": "^0.2.3",
-		"style-loader": "^0.20.3",
-		"typescript": "^3.7.5",
+		"rimraf": "^3.0.2",
+		"source-map-loader": "^3.0.1",
+		"style-loader": "^3.3.1",
+		"typescript": "^4.5.5",
 		"webpack-cli": "^4.9.2",
-		"webpack-merge": "^4.1.2"
+		"webpack-merge": "^5.8.0"
 	}
 }
